refactor: move project into dedicated folder
