+++ lib/health-check/health-check-result.interface.ts
@@ -1,0 +1,7 @@
+/**
+ * The result of a health check
+ */
+export interface HealthCheckResult {
+  results: any[];
+  errors: any[];
+}
