+++ lib/health-check/index.ts
@@ -1,0 +1,3 @@
+export * from './health-check.service';
+export * from './health-check.decorator';
+export * from './health-check-result.interface';
