+++ lib/health-indicator/database/databse-ping-check-settings.interface.ts
