+++ lib/health-indicator/disk/disk.health.ts
@@ -2,8 +2,7 @@ import { Injectable, Inject } from '@nestjs/common';
 import { isNil } from '@nestjs/common/utils/shared.utils';
 import * as checkDiskSpace from 'check-disk-space';
 
-import { HealthIndicatorResult } from '../../interfaces';
-import { HealthIndicator } from '../health-indicator';
+import { HealthIndicator, HealthIndicatorResult } from '../health-indicator';
 import { CHECKDISKSPACE_LIB } from '../../terminus.constants';
 import { StorageExceededError } from '../../errors';
 import { STORAGE_EXCEEDED } from '../../errors/messages.constant';
