+++ lib/health-indicator/dns/index.ts
@@ -1,0 +1,1 @@
+export * from './dns.health';
