+++ lib/health-indicator/microservice/protos/health.proto
