+++ lib/index.ts
@@ -1,4 +1,5 @@
-export * from './interfaces';
-export * from './terminus.module';
-export * from './health-indicators';
+export { TerminusModule } from './terminus.module';
+export * from './health-indicator';
 export * from './errors';
+export { HealthCheck, HealthCheckService } from './health-check';
+export * from './terminus-module-options.interface';
