--- lib/interfaces/index.ts
@@ -1,2 +1,0 @@
-export * from './terminus-module-options.interface';
-export * from './health-indicator.interface';
